creating and downloading isdoc

+++ project/blueprints/site.py
@@ -1,5 +1,6 @@
+import io
 import os
-from flask import Blueprint, render_template, request, flash, redirect
+from flask import Blueprint, render_template, request, flash, redirect, send_file
 from flask_login import login_required, login_user, current_user, logout_user
 from project.models import User
 from project.database import mongo
@@ -9,8 +10,6 @@ site = Blueprint('site', __name__)
 
 @site.route('/', methods=['GET', 'POST'])
 def login():
-    ''' login for authenticated and not authenticated user '''
-
     # Cannot access member "is_authenticated" for type "LocalProxy"
     if current_user.is_authenticated and request.method == 'GET': # type: ignore 
         return redirect('/home')
@@ -40,11 +39,17 @@ def login():
 
     return render_template('login.html'), 200
 
+
+@site.route('/logout', methods=['GET'])
+@login_required
+def logout():
+    logout_user()
+    return redirect('/')
+
+
 @site.route('/home', methods=['GET', 'POST'])
 @login_required
 def home():
-    ''' display home page to user or generate isdoc file from submitted PDF and allow him/her to download it '''
-
     if request.method == 'POST':
         pdf_file = request.files["pdf_file"]
 
@@ -56,17 +61,43 @@ def home():
 
         pdf_file.save(pdf_file_name)
 
-        # TODO: create isdoc
-        # pdf_parser.create_is_doc()
+        try:
+            invoice_object = pdf_parser.parse_pdf(file_name=pdf_file_name)
 
-        os.remove(pdf_file_name)
+            os.remove(pdf_file_name)
 
-        return render_template('download.html'), 200
+            if invoice_object.success_parse_flag is False:
+                for err in invoice_object.parsing_errors:
+                    flash(err, category='error')
+                
+                return render_template('home.html'), 400
+
+            creation_errors = pdf_parser.create_is_doc(invoice_object)
+
+            if len(creation_errors):
+                for err in creation_errors:
+                    flash(err, category='err')
+
+                return render_template('home.html'), 400
+
+            return_data = io.BytesIO()
+            isdoc_file_name = invoice_object.isdoc_name()
+            
+            with open(isdoc_file_name, 'rb') as file:
+                return_data.write(file.read())
+
+            return_data.seek(0) # return cursor from last byte to first one
+
+            os.remove(isdoc_file_name)
+
+            return send_file(return_data, mimetype='text/xml',
+             as_attachment=True, attachment_filename=isdoc_file_name)
+
+        except:
+            if os.path.isfile(pdf_file_name):
+                os.remove(pdf_file_name)
+
+            flash('Aplikácia narazila na problém, kontaktujte správcu.', category='error')
+            return render_template('home.html'), 400
 
     return render_template('home.html'), 200
-
-@site.route('/logout', methods=['GET'])
-@login_required
-def logout():
-    logout_user()
-    return redirect('/')
